% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.osmtile}
\alias{autoplot.osmtile}
\title{Plots an open street map tile using ggplot2}
\usage{
\method{autoplot}{osmtile}(object, plot = FALSE, ...)
}
\arguments{
\item{object}{an osmtile}

\item{plot}{if false only the annotation_raster is returned}

\item{...}{not used}
}
\description{
Plots an open street map tile using ggplot2
}
