% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot}
\alias{autoplot}
\title{Create a complete ggplot appropriate to a particular data type
Rexported from ggplot2. `autoplot()` uses ggplot2 to draw a particular plot for an object of a
particular class in a single command. This defines the S3 generic that
other classes and packages can extend.}
\arguments{
\item{object}{An object (for example an OpenStreetMap object)}

\item{...}{Additional arguments}
}
\value{
a ggplot object
}
\description{
Create a complete ggplot appropriate to a particular data type
Rexported from ggplot2. `autoplot()` uses ggplot2 to draw a particular plot for an object of a
particular class in a single command. This defines the S3 generic that
other classes and packages can extend.
}
\seealso{
[autoplot.OpenStreetMap()]
}
