% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osm.R
\name{getMapInfo}
\alias{getMapInfo}
\title{Returns a table with relevant source and attribution info for each map type}
\usage{
getMapInfo()
}
\description{
Returns a table with relevant source and attribution info for each map type
}

