\name{openproj.default}
\alias{openproj.default}
\title{Projects the open street map to an alternate coordinate system}
\usage{
  openproj.default(x, projection = "+proj=longlat", ...)
}
\arguments{
  \item{x}{an OpenStreetMap object}

  \item{projection}{a proj4 character string or CRS object}

  \item{...}{additional parameters for projectRaster}
}
\description{
  Projects the open street map to an alternate coordinate
  system
}

