\docType{package}
\name{OpenStreetMap-package}
\alias{OpenStreetMap}
\alias{OpenStreetMap-package}
\title{Open Street Maps.}
\description{
  Open street maps.
}
\details{
  \tabular{ll}{ Package: \tab OpenStreetMap\cr Type: \tab
  Package\cr Version: \tab 1.0\cr Date: \tab 2011-04-04\cr
  License: \tab LGPL-2\cr LazyLoad: \tab yes\cr }

  Includes The ability to plot Open Street Map and Bing
  satellite images.
}
\author{
  Ian Fellows \email{ian.fellows@stat.ucla.edu}
}
\references{
  \url{http://www.fellstat.com}
}
\keyword{package}

