% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osm.R
\name{plot.osmtile}
\alias{plot.osmtile}
\title{Add tile to plot}
\usage{
\method{plot}{osmtile}(x, y = NULL, add = TRUE, raster = TRUE, ...)
}
\arguments{
\item{x}{the tile}

\item{y}{ignored}

\item{add}{add to current plot (if raster, then image is always added)}

\item{raster}{use raster image}

\item{...}{additional parameters to image or rasterImage}
}
\description{
Add tile to plot
}

