\name{project_mercator}
\alias{project_mercator}
\title{maps long lat values to the open street map mercator projection}
\usage{
  project_mercator(lat, long, drop = TRUE)
}
\arguments{
  \item{lat}{a vector of latitudes}

  \item{long}{a vector of longitudes}

  \item{drop}{drop to lowest dimension}
}
\description{
  maps long lat values to the open street map mercator
  projection
}

