% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osm.R
\name{projectMercator}
\alias{projectMercator}
\title{Maps long lat values to the open street map mercator projection}
\usage{
projectMercator(lat, long, drop = TRUE)
}
\arguments{
\item{lat}{a vector of latitudes}

\item{long}{a vector of longitudes}

\item{drop}{drop to lowest dimension}
}
\description{
Maps long lat values to the open street map mercator projection
}

