% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osm.R
\docType{methods}
\name{raster,osmtile-method}
\alias{raster,osmtile-method}
\title{Create a RasterLayer from a tile}
\usage{
\S4method{raster}{osmtile}(x, ...)
}
\arguments{
\item{x}{an osmtile}

\item{...}{unused}
}
\description{
Create a RasterLayer from a tile
}

