% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/osm.R
\name{setCloudMadeKey}
\alias{setCloudMadeKey}
\title{Sets the user identification key for cloudmade.com}
\usage{
setCloudMadeKey(key)
}
\arguments{
\item{key}{The key. Obtain a (not-)free map key at http://www.cloudmade.com}
}
\description{
Sets the user identification key for cloudmade.com
}

