\name{autoplot.osmtile}
\alias{autoplot.osmtile}
\title{Plots an open street map tile using ggplot2}
\usage{
  \method{autoplot}{osmtile} (data, plot = FALSE, ...)
}
\arguments{
  \item{data}{an osmtile}

  \item{plot}{if false only the annotation_raster is
  returned}

  \item{...}{not used}
}
\description{
  Plots an open street map tile using ggplot2
}

