\name{openmap}
\alias{openmap}
\title{get a map based on lat long coordinates}
\usage{
  openmap(upperLeft, lowerRight, zoom, type = "osm")
}
\arguments{
  \item{upperLeft}{the upper left lat and long}

  \item{lowerRight}{the lower right lat and long}

  \item{zoom}{the zoom level}

  \item{type}{osm for mapnik open street map, or 'bing' for
  bing aerial}
}
\description{
  get a map based on lat long coordinates
}

