\name{osmtile}
\alias{osmtile}
\title{Get an open street map tile.}
\usage{
  osmtile(x, y, zoom, type = "osm")
}
\arguments{
  \item{x}{location in osm native coordinates}

  \item{y}{location in osm native coordinates}

  \item{zoom}{zoom level}

  \item{type}{the map type (see getMapInfo)}
}
\value{
  a tile
}
\description{
  Get an open street map tile.
}

