% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm.R
\name{plot.OpenStreetMap}
\alias{plot.OpenStreetMap}
\title{Plot an OpenStreetMap object.}
\usage{
\method{plot}{OpenStreetMap}(x, y = NULL, add = FALSE, removeMargin = TRUE, ...)
}
\arguments{
\item{x}{the OpenStreetMap}

\item{y}{ignored}

\item{add}{add to current plot}

\item{removeMargin}{remove margins from plotting device}

\item{...}{additional parameters to be passed to plot}
}
\description{
Plot an OpenStreetMap object.
}
