\name{plot.osmtile}
\alias{plot.osmtile}
\title{add tile to plot}
\usage{
  \method{plot}{osmtile} (x, y = NULL, add = TRUE,
    raster = FALSE, ...)
}
\arguments{
  \item{x}{the tile}

  \item{y}{ignored}

  \item{add}{add to current plot (if raster, then image is
  always added)}

  \item{raster}{use raster image}

  \item{...}{additional parameters to image or rasterImage}
}
\description{
  add tile to plot
}

